Move `impl_binary_value_for_message` to proto module.
