+++ exonum/src/node/mod.rs
@@ -1121,9 +1121,7 @@ mod tests {
     use crate::events::EventHandler;
     use crate::helpers;
     use crate::proto::{schema::tests::TxSimple, ProtobufConvert};
-    use exonum_merkledb::{
-        impl_binary_value_for_message, BinaryValue, Database, Snapshot, TemporaryDB,
-    };
+    use exonum_merkledb::{BinaryValue, Database, Snapshot, TemporaryDB};
     use protobuf::Message as ProtobufMessage;
 
     const SERVICE_ID: u16 = 0;
