+++ exonum/src/sandbox/config_updater.rs
@@ -24,7 +24,7 @@ use crate::crypto::{Hash, PublicKey, SecretKey};
 use crate::helpers::Height;
 use crate::messages::{Message, RawTransaction, Signed};
 use crate::proto::ProtobufConvert;
-use exonum_merkledb::{impl_binary_value_for_message, BinaryValue, Snapshot};
+use exonum_merkledb::{BinaryValue, Snapshot};
 use protobuf::Message as PbMessage;
 
 pub const CONFIG_SERVICE: u16 = 1;
