+++ exonum/src/sandbox/timestamping.rs
@@ -23,7 +23,7 @@ use crate::blockchain::{
 };
 use crate::crypto::{gen_keypair, Hash, PublicKey, SecretKey, HASH_SIZE};
 use crate::messages::{Message, RawTransaction, Signed};
-use exonum_merkledb::{impl_binary_value_for_message, BinaryValue, Snapshot};
+use exonum_merkledb::{BinaryValue, Snapshot};
 use protobuf::Message as PbMessage;
 
 pub const TIMESTAMPING_SERVICE: u16 = 129;
